+++ packages/backend-ai/src/conversation/index.ts
@@ -1,0 +1,9 @@
+import {
+  DefineConversationHandlerFunctionProps,
+  defineConversationHandlerFunction,
+} from './factory.js';
+
+export {
+  DefineConversationHandlerFunctionProps,
+  defineConversationHandlerFunction,
+};