+++ packages/platform-core/src/caller_directory_extractor.test.ts
@@ -1,18 +1,20 @@
 import { describe, it } from 'node:test';
 import assert from 'node:assert';
-import { getCallerDirectory } from './get_caller_directory.js';
+import { CallerDirectoryExtractor } from './caller_directory_extractor.js';
 import { fileURLToPath, pathToFileURL } from 'url';
 
 void describe('getCallerDirectory', () => {
   void it('throws if stack trace is undefined', () => {
-    assert.throws(() => getCallerDirectory(undefined));
+    assert.throws(() => new CallerDirectoryExtractor(undefined).extract());
   });
 
   void describe('typescript app', () => {
     void it('throws if stack trace has less than two frames', () => {
       const tooShortStack = `Error
     at AmplifyAuthFactory (/Users/alias/sandboxes/install-test/node_modules/@aws-amplify/backend-auth/src/factory.ts:28:24)`;
-      assert.throws(() => getCallerDirectory(tooShortStack));
+      assert.throws(() =>
+        new CallerDirectoryExtractor(tooShortStack).extract()
+      );
     });
 
     void it('throws if regex match not found', () => {
@@ -20,7 +22,9 @@ void describe('getCallerDirectory', () => {
     at AmplifyAuthFactory some garbage
     at more garbage
     at Object.<anonymous> (/Users/alias/sandboxes/install-test/backend/auth.ts:5:2)`;
-      assert.throws(() => getCallerDirectory(malformedStacktrace));
+      assert.throws(() =>
+        new CallerDirectoryExtractor(malformedStacktrace).extract()
+      );
     });
 
     void it('returns path of second stack frame', () => {
@@ -29,7 +33,7 @@ void describe('getCallerDirectory', () => {
     at <anonymous> (/Users/alias/sandboxes/install-test/backend/otherName.ts:3:21)
     at Object.<anonymous> (/Users/alias/sandboxes/install-test/backend/auth.ts:5:2)`;
       assert.strictEqual(
-        getCallerDirectory(validStackTrace),
+        new CallerDirectoryExtractor(validStackTrace).extract(),
         '/Users/alias/sandboxes/install-test/backend'
       );
     });
@@ -39,7 +43,9 @@ void describe('getCallerDirectory', () => {
     void it('throws if stack trace has less than two frames', () => {
       const tooShortStack = `Error
     at file:///Users/alias/sandboxes/install-test/node_modules/@aws-amplify/backend-auth/src/factory.ts:28:24`;
-      assert.throws(() => getCallerDirectory(tooShortStack));
+      assert.throws(() =>
+        new CallerDirectoryExtractor(tooShortStack).extract()
+      );
     });
 
     void it('throws if regex match not found', () => {
@@ -47,7 +53,9 @@ void describe('getCallerDirectory', () => {
     at AmplifyAuthFactory some garbage
     at more garbage
     at file:///Users/alias/sandboxes/install-test/backend/auth.ts:5:2`;
-      assert.throws(() => getCallerDirectory(malformedStacktrace));
+      assert.throws(() =>
+        new CallerDirectoryExtractor(malformedStacktrace).extract()
+      );
     });
 
     void it('returns path of second stack frame', () => {
@@ -57,7 +65,7 @@ void describe('getCallerDirectory', () => {
     at ${someFileUrl.toString()}/otherName.ts:3:21
     at file:///Users/alias/sandboxes/install-test/backend/auth.ts:5:2`;
       assert.strictEqual(
-        getCallerDirectory(validStackTrace),
+        new CallerDirectoryExtractor(validStackTrace).extract(),
         fileURLToPath(someFileUrl)
       );
     });