+++ packages/backend-secret/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @aws-amplify/backend-secret
 
+## 1.1.1
+
+### Patch Changes
+
+- 79fdc7c: wrap InvalidSignatureException in AmplifyUserError
+
 ## 1.1.0
 
 ### Minor Changes